added score and adjusted currency

+++ include/entities/entities.h
@@ -25,7 +25,7 @@ typedef struct Entities
     void (*add_ball)(Player *player, b2WorldId world_id, Paddle *paddle);
     void (*add_paddle)(Player *player, b2WorldId world_id);
     void (*add_brick)(struct GameContext *game_context, b2Vec2 position, int brick_type);
-    void (*add_nanite)(b2WorldId world_id, b2Vec2 position, int currency);
+    void (*add_nanite)(b2WorldId world_id, b2Vec2 position, float currency);
     void (*add_wall_edges)(b2WorldId world_id);
     void (*add_kill_boundary)(b2WorldId world_id);
     void (*update)(float delta_time);
