+++ include/entities/nanite.h
@@ -16,15 +16,15 @@ typedef struct Nanite
 
     Texture2D *texture;
     b2Vec2 size;
-    int currency;
+    float currency;
 
     void (*update)(struct Nanite *nanite, float delta_time);
     void (*render)(struct Nanite *nanite);
     void (*clean_up)(struct Nanite *nanite);
-    void (*reset)(struct Nanite *nanite, b2Vec2 position, int currency);
+    void (*reset)(struct Nanite *nanite, b2Vec2 position, float currency);
     void (*disable)(struct Nanite *nanite);
 } Nanite;
 
-Nanite *create_nanite(b2WorldId world_id, b2Vec2 position, int currency);
+Nanite *create_nanite(b2WorldId world_id, b2Vec2 position, float currency);
 
 #endif
